requirements.txt
streamlit
mysql-connector-python
pandas
psycopg2-binary
python-dotenv
